+++ app/src/test/java/org/simple/clinic/home/overdue/OverdueUpdateTest.kt
@@ -20,7 +20,9 @@ import java.util.UUID
 class OverdueUpdateTest {
 
   private val dateOnClock = LocalDate.parse("2018-01-01")
-  private val updateSpec = UpdateSpec(OverdueUpdate(date = dateOnClock, canGeneratePdf = true))
+  private val updateSpec = UpdateSpec(OverdueUpdate(date = dateOnClock,
+      canGeneratePdf = true,
+      isOverdueSectionsFeatureEnabled = false))
   private val defaultModel = OverdueModel.create()
 
   @Test
@@ -62,7 +64,7 @@ class OverdueUpdateTest {
   }
 
   @Test
-  fun `when current facility is loaded and overdue list changes feature is enabled, then load overdue appointments with patients with no phone numbers`() {
+  fun `when current facility is loaded and overdue sections feature is disabled, then load pending overdue appointments`() {
     val facility = TestData.facility(
         uuid = UUID.fromString("6d66fda7-7ca6-4431-ac3b-b570f1123624"),
         facilityConfig = FacilityConfig(
@@ -77,6 +79,31 @@ class OverdueUpdateTest {
         .then(assertThatNext(
             hasModel(defaultModel.currentFacilityLoaded(facility)),
             hasEffects(LoadOverdueAppointments_old(dateOnClock, facility))
+        ))
+  }
+
+  @Test
+  fun `when current facility is loaded and overdue sections feature is enabled, then load overdue appointments`() {
+    val facility = TestData.facility(
+        uuid = UUID.fromString("6d66fda7-7ca6-4431-ac3b-b570f1123624"),
+        facilityConfig = FacilityConfig(
+            diabetesManagementEnabled = true,
+            teleconsultationEnabled = false
+        )
+    )
+
+    val updateSpec = UpdateSpec(OverdueUpdate(
+        date = dateOnClock,
+        canGeneratePdf = true,
+        isOverdueSectionsFeatureEnabled = true
+    ))
+
+    updateSpec
+        .given(defaultModel)
+        .whenEvent(CurrentFacilityLoaded(facility))
+        .then(assertThatNext(
+            hasModel(defaultModel.currentFacilityLoaded(facility)),
+            hasEffects(LoadOverdueAppointments(dateOnClock, facility))
         ))
   }
 
@@ -104,7 +131,7 @@ class OverdueUpdateTest {
 
   @Test
   fun `when download overdue list button is clicked, network is connected and pdf can not be generated, then schedule download`() {
-    val updateSpec = UpdateSpec(OverdueUpdate(date = dateOnClock, canGeneratePdf = false))
+    val updateSpec = UpdateSpec(OverdueUpdate(date = dateOnClock, canGeneratePdf = false, isOverdueSectionsFeatureEnabled = false))
 
     updateSpec
         .given(defaultModel)
@@ -139,7 +166,7 @@ class OverdueUpdateTest {
 
   @Test
   fun `when share overdue list button is clicked, network is connected but pdf can not be generated, then open progress for share dialog`() {
-    val updateSpec = UpdateSpec(OverdueUpdate(date = dateOnClock, canGeneratePdf = false))
+    val updateSpec = UpdateSpec(OverdueUpdate(date = dateOnClock, canGeneratePdf = false, isOverdueSectionsFeatureEnabled = false))
 
     updateSpec
         .given(defaultModel)